// packages/shared/src/noise/quantiles.generated.ts
export const QUANTILE_PROBABILITIES: readonly number[] = Object.freeze([
  0, 0.001, 0.01, 0.025, 0.05, 0.1, 0.2, 0.3, 0.4, 0.5, 0.55, 0.6, 0.7, 0.8,
  0.9, 0.95, 0.975, 0.99, 0.999, 1,
]);
export const NOISE_QUANTILES = Object.freeze({
  openSimplex2: Object.freeze([
    -0.9999555406433706, -0.9788188717833524, -0.9238170270542675,
    -0.8879371121679012, -0.8410387589456573, -0.7485680313405301,
    -0.5643946801559847, -0.3770513328615949, -0.18856844024257852,
    0.0004647378757170462, 0.09456808604865653, 0.18955924159391016,
    0.3777657817233779, 0.5653350339787151, 0.7499966375721009,
    0.841995544281413, 0.8883520908390231, 0.924036539926203,
    0.9780453242300177, 0.9999854712767007,
  ]),
  openSimplex3: Object.freeze([
    -0.9994831119204023, -0.9736022260049682, -0.8978327042130957,
    -0.8176460676542938, -0.7189826945769123, -0.5772381731904986,
    -0.40420379435724324, -0.26117260303565376, -0.1291179063686563,
    0.0000016324096417710786, 0.06382560990461861, 0.12921901841013153,
    0.2610320669170731, 0.4039292606090296, 0.5774373933086586,
    0.719197091776222, 0.8189278062958345, 0.8978560604407837,
    0.9745636914559589, 0.9993871839794455,
  ]),
  openSimplex2S3: Object.freeze([
    -0.9987912370444805, -0.9266333252898202, -0.8003254469612237,
    -0.6979287265506934, -0.6163432293011449, -0.5169914684598815,
    -0.37069514391568775, -0.23970722700738414, -0.1172238833962403,
    0.00028600345840043513, 0.05809992855998804, 0.11748628292279024,
    0.24069785314843056, 0.37158501816647194, 0.5181165702094023,
    0.6173871964111424, 0.6981280227908648, 0.8013512603993246,
    0.9282497508760608, 0.9954402197055727,
  ]),
  fbm2: Object.freeze([
    -0.9334757957518568, -0.7736633930939216, -0.66205019596541,
    -0.5934262912197483, -0.5240243315290831, -0.43190593876953653,
    -0.29978012432537887, -0.19221724538114826, -0.09476141906158983,
    -0.00028279217521277563, 0.0466609935846138, 0.09412803971102475,
    0.19217599562717258, 0.29938808230014985, 0.431768682210393,
    0.5253854064315994, 0.5952265483502664, 0.6635961153320227,
    0.7747776029184571, 0.9501676340240558,
  ]),
  ridged2: Object.freeze([
    -0.999999999652505, -0.9997716608616256, -0.9957183737371228,
    -0.9894780088801243, -0.980906494282263, -0.9619070800977876,
    -0.9072374925699448, -0.8278946593430254, -0.7217275422021829,
    -0.5890523406168545, -0.511493333027833, -0.42798440531553833,
    -0.2502001065285583, -0.06539483818327525, 0.13702584254262432,
    0.3001517678200614, 0.42245479617381576, 0.5443814008816932,
    0.7473515151468646, 0.9954544326279924,
  ]),
  billow2: Object.freeze([
    -0.9977239683333022, -0.8584146281806173, -0.7392514850583602,
    -0.666157164403405, -0.5926608982301792, -0.49552116192284323,
    -0.3613985532783884, -0.2536271778372476, -0.15366882263350487,
    -0.05477741951097792, -0.0059249034125643234, 0.042350051123083626,
    0.13877768585229602, 0.2406910605580623, 0.3665339892335294,
    0.45811759965123683, 0.5284553861120443, 0.5988594986697644,
    0.7142263862626248, 0.9003352680481117,
  ]),
  erosionFbm2: Object.freeze([
    -0.7514212701073254, -0.4798036488136239, -0.3330284653243553,
    -0.22459974393371746, -0.14348495309347944, -0.0781570741976886,
    -0.03310371482435533, -0.015813249205979228, -0.006469006772729564,
    -0.000027578025917615153, 0.0030144673665930912, 0.006429535082782504,
    0.01574814855945916, 0.033091159922732497, 0.07833006280303072,
    0.14418272563104606, 0.22483324299806223, 0.3322379434848077,
    0.4773182388040642, 0.721977900409245,
  ]),
  warp2Component: Object.freeze([
    -0.9817160957794162, -0.9012733010801066, -0.8044755595088268,
    -0.7360727317178176, -0.6586636205006103, -0.5452349879498999,
    -0.3790337998048361, -0.24616216131751956, -0.12099264985142712,
    0.00022533645792860435, 0.05926577183019378, 0.12047338067643655,
    0.2458706420053844, 0.37903029053765974, 0.5452110147213514,
    0.6600052864973585, 0.7373594256671246, 0.8065851861719711,
    0.9003283880565072, 0.9758025407865812,
  ]),
  warp3Component: Object.freeze([
    -0.9324709573707214, -0.7621851068768933, -0.6206165004365461,
    -0.5431537913033994, -0.4718294403971578, -0.3823475677474923,
    -0.26151449172288266, -0.16761697171475254, -0.08262501479751663,
    -0.00002465354323027159, 0.040666149954348994, 0.0825264722555848,
    0.16806679953033501, 0.2620718814357027, 0.3826305163538466,
    0.4727739047478718, 0.5437525649553818, 0.6226585285228246,
    0.7642753868572105, 0.9301027673900845,
  ]),
});
export type NoiseDistribution = keyof typeof NOISE_QUANTILES;